Added a build that runs on the Raspberry Pi 32-bit
This has been such an interesting project to contribute to... :)

So idk what -DUNALIGNED_DOUBLES=1 is, just like I didn't know what that crazy old GC was (apparently now it comes standard with the build-essential or glibc packages?).  But on my Raspberry Pi 3B+, which I think is 32-bit, if I use that flag I end up with a binary that only works in direct mode.  If I try to enter a program, it crashed with "bus error" and some other stuff about "alignment".  I'm guessing it's somthing suuuuuper low-level, like to the point of how doubles are implemented at the kernel level... but regardless, removing the flag fixes the problem, and fbasic now works on old models of Raspberry Pi! :)